+++ cookbooks/arcgis-desktop/providers/licensemanager.rb
@@ -16,6 +16,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+use_inline_resources if defined?(use_inline_resources)
+
 action :system do
   if node['platform'] == 'windows'
     #TODO: Ensure License Manager system requirements on Windows