+++ cookbooks/esri-tomcat/attributes/default.rb
@@ -7,5 +7,10 @@ default['tomcat']['group'] = 'tomcat_' + node['tomcat']['instance_name']
 
 default['tomcat']['ssl_enabled_protocols']='TLSv1.2,TLSv1.1,TLSv1'
 default['tomcat']['keystore_file']  = ''
-default['tomcat']['keystore_password']  = ''
+if ENV['TOMCAT_KEYSTORE_PASSWORD'].nil?
+  default['tomcat']['keystore_password']  = ''
+else
+  default['tomcat']['keystore_password']  = ENV['TOMCAT_KEYSTORE_PASSWORD']
+end
+default['tomcat']['keystore_type']  = 'PKCS12'
 default['tomcat']['domain_name']  = node['fqdn']